refactor: modularize task scheduler
